+++ src/main/java/org/terasology/launcher/game/GameStarter.java
@@ -1,12 +1,14 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
 package org.terasology.launcher.game;
 
+import com.vdurmont.semver4j.Semver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.event.Level;
 import org.terasology.launcher.util.JavaHeapSize;
+import org.terasology.launcher.util.Platform;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -24,9 +26,10 @@ class GameStarter implements Callable<Process> {
     private static final Logger logger = LoggerFactory.getLogger(GameStarter.class);
 
     final ProcessBuilder processBuilder;
+    private final Semver engineVersion;
 
     /**
-     * @param gamePath          the directory under which we will find {@code libs/Terasology.jar}, also used as the process's
+     * @param installation          the directory under which we will find {@code libs/Terasology.jar}, also used as the process's
      *                          working directory
      * @param gameDataDirectory {@code -homedir}, the directory where Terasology's data files (saves & etc) are kept
      * @param heapMin           java's {@code -Xms}
@@ -35,12 +38,14 @@ class GameStarter implements Callable<Process> {
      * @param gameParams        additional arguments for the Terasology command line
      * @param logLevel          the minimum level of log events Terasology will include on its output stream to us
      */
-    GameStarter(Path gamePath, Path gameDataDirectory,
-                JavaHeapSize heapMin, JavaHeapSize heapMax,
-                List<String> javaParams,
-                List<String> gameParams,
-                Level logLevel) {
+    GameStarter(Installation installation, Path gameDataDirectory, JavaHeapSize heapMin, JavaHeapSize heapMax,
+                List<String> javaParams, List<String> gameParams, Level logLevel) throws IOException {
+        engineVersion = installation.getEngineVersion();
+        var gamePath = installation.path;
+
+        final boolean isMac = Platform.getPlatform().isMac();
         final List<String> processParameters = new ArrayList<>();
+
         processParameters.add(getRuntimePath().toString());
 
         if (heapMin.isUsed()) {
@@ -50,6 +55,14 @@ class GameStarter implements Callable<Process> {
             processParameters.add("-Xmx" + heapMax.getSizeParameter());
         }
         processParameters.add("-DlogOverrideLevel=" + logLevel.name());
+
+        if (isMac && VersionHistory.LWJGL3.isProvidedBy(engineVersion)) {
+            processParameters.add("-XstartOnFirstThread");  // lwjgl3 requires this on OS X
+            // awt didn't work either, but maybe fixed on newer versions?
+            //   https://github.com/LWJGLX/lwjgl3-awt/issues/1
+            processParameters.add("-Djava.awt.headless=true");
+        }
+
         processParameters.addAll(javaParams);
 
         // Locate the main game jar. Currently, Terasology has custom build logic to put libraries into a 'libs'
@@ -57,6 +70,7 @@ class GameStarter implements Callable<Process> {
         // how the game distribution was build (i.e., custom 'libs' or default 'lib').
         //TODO: this should probably be part of ReleaseMetadata and be determined further up the hierarchy
         processParameters.add("-jar");
+      
         if (gamePath.resolve("libs").toFile().isDirectory()) {
             // custom Terasology build logic puts libraries into 'libs' subdirectory
             processParameters.add(gamePath.resolve(Path.of("libs", "Terasology.jar")).toString());
@@ -65,8 +79,14 @@ class GameStarter implements Callable<Process> {
             processParameters.add(gamePath.resolve(Path.of("lib", "Terasology.jar")).toString());
         }
 
-        processParameters.add("-homedir=" + gameDataDirectory.toAbsolutePath().toString());
+        // Parameters after this are for the game facade, not the java runtime.
+        processParameters.add(homeDirParameter(gameDataDirectory));
         processParameters.addAll(gameParams);
+
+        if (isMac) {
+            // splash screen uses awt, so no awt => no splash
+            processParameters.add(noSplashParameter());
+        }
 
         processBuilder = new ProcessBuilder(processParameters)
                 .directory(gamePath.toFile())
@@ -90,5 +110,21 @@ class GameStarter implements Callable<Process> {
      */
     Path getRuntimePath() {
         return Paths.get(System.getProperty("java.home"), "bin", "java");
+    }
+
+    String homeDirParameter(Path gameDataDirectory) {
+        if (terasologyUsesPosixOptions()) {
+            return "--homedir=" + gameDataDirectory.toAbsolutePath();
+        } else {
+            return "-homedir=" + gameDataDirectory.toAbsolutePath();
+        }
+    }
+
+    String noSplashParameter() {
+        return terasologyUsesPosixOptions() ? "--no-splash" : "-noSplash";
+    }
+
+    boolean terasologyUsesPosixOptions() {
+        return VersionHistory.PICOCLI.isProvidedBy(engineVersion);
     }
 }