+++ src/main/java/org/terasology/launcher/model/ReleaseMetadata.java
@@ -4,29 +4,26 @@
 package org.terasology.launcher.model;
 
 import java.util.Date;
-import java.util.List;
 
 /**
  * Data container for metadata associated with a game release.
  *
- * The metadata in this class is either relevant for displaying more information to the user (e.g., {@code changelog},
- * {@code timestamp}) or for managing and starting the game itself (e.g., {@code isLwjgl3}).
+ * The metadata in this class is relevant for displaying more information to the user, e.g., {@code changelog},
+ * {@code timestamp}.
  */
 public class ReleaseMetadata {
-    private final List<String> changelog;
+    private final String changelog;
     private final Date timestamp;
-    private final boolean isLwjgl3;
 
-    public ReleaseMetadata(List<String> changelog, Date timestamp, boolean isLwjgl3) {
+    public ReleaseMetadata(String changelog, Date timestamp) {
         this.changelog = changelog;
         this.timestamp = timestamp;
-        this.isLwjgl3 = isLwjgl3;
     }
 
     /**
-     * The change log of this release, each line represents one logical change.
+     * The change log of this release as a single markdown string.
      */
-    public List<String> getChangelog() {
+    public String getChangelog() {
         return changelog;
     }
 
@@ -35,12 +32,5 @@ public class ReleaseMetadata {
      */
     public Date getTimestamp() {
         return timestamp;
-    }
-
-    /**
-     * Whether this release uses LWJGL v3 or not.
-     */
-    public boolean isLwjgl3() {
-        return isLwjgl3;
     }
 }